requirements.txt
anthropic==0.25.1
docx2txt==0.8
langchain==0.1.16
langchain-anthropic==0.1.8
langchain-community==0.0.32
langchain-experimental==0.0.57
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
openai==1.17.1
pinecone-client==3.2.2
pdfminer.six==20231228
streamlit